+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 - `defaultObjectBySchema()` improved support for `ZodDefault` and `ZodArray`;
 - Dependencies update.
 
+### Added
+
+- Replace `vue-tsc` with `vite-plugin-dts` for types generation.
+
 ## [0.0.13] - 2023-05-19
 
 ### Fixed
@@ -16,7 +20,7 @@ All notable changes to this project will be documented in this file.
 - `VvFormTemplate` and `VvFormField` support for `ref()` props;
 - `VvFormField` datetime correct type is `datetime-local`.
 
-### Updated
+### Added
 
 - `@volverjs/ui-vue` to `v0.0.8-beta.4` and added to peerDependencies